requirements.txt
langgraph>=1.0.0
langchain>=0.3.0
langchain-google-genai>=4.1.1
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
pypdf>=4.0.0
docx2txt>=0.9
unstructured>=0.15.0
ruff>=0.4.0
